add direction

+++ inc/figure.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   figure.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkulikov <mkulikov@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: alvutina <alvutina@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 11:36:18 by mkulikov          #+#    #+#             */
-/*   Updated: 2024/12/28 21:08:07 by mkulikov         ###   ########.fr       */
+/*   Updated: 2025/01/14 11:19:09 by alvutina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ t_figure	*get_plane(t_ftype type, char **s);
 t_figure	*get_sphere(t_ftype type, char **s);
 t_ftype		get_ftype(char *id);
 float			sphere_intersect(t_camera *camera, t_vector *ray, t_figure *sphere);
-int			plane_intersect(t_camera *camera, t_vector *ray, t_figure *plane);
-int			cylinder_intersect(t_camera *camera, t_vector *ray, t_figure *plane);
+float			plane_intersect(t_camera *camera, t_vector *ray, t_figure *plane);
+float			cylinder_intersect(t_camera *camera, t_vector *ray, t_figure *plane);
 int			get_figure_color(t_figure *figure, t_vector *p, float t, t_data *d);
 
 #endif
